Added reviewer role applications
Created a file upload for reviewer role applications.  It uses the same code as the file upload of the other files, but the document type is locked into the value "Reviewer".  I also updated the logic so this doesn't show up in the My Files page because the no one but the admins should be able to see it.  Speaking of admins, I also updated the admins page with the My Page code, except it only shows documents of the "Reviewer" type.  I could not test it since I could not access the page, but I would guess either the page is just broken, or it works.